update zlib-ng: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@65900 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/zlib/arch/loongarch/chunkset_lasx.c
@@ -2,16 +2,18 @@
  * Copyright (C) 2025 Vladislav Shchapov <vladislav@shchapov.ru>
  * For conditions of distribution and use, see copyright notice in zlib.h
  */
-#include "zbuild.h"
-#include "zmemory.h"
 
 #ifdef LOONGARCH_LASX
+
+#include "zbuild.h"
+#include "zsanitizer.h"
+#include "zmemory.h"
 
 #include <lasxintrin.h>
 #include "lasxintrin_ext.h"
 #include "lsxintrin_ext.h"
 
-#include "arch/generic/chunk_256bit_perm_idx_lut.h"
+#include "arch/shared/chunk_256bit_perm_idx_lut.h"
 
 typedef __m256i chunk_t;
 typedef __m128i halfchunk_t;
@@ -47,7 +49,7 @@ static inline void storechunk(uint8_t *out, chunk_t *chunk) {
     __lasx_xvst(*chunk, out, 0);
 }
 
-static inline chunk_t GET_CHUNK_MAG(uint8_t *buf, uint32_t *chunk_rem, uint32_t dist) {
+static inline chunk_t GET_CHUNK_MAG(uint8_t *buf, size_t *chunk_rem, size_t dist) {
     lut_rem_pair lut_rem = perm_idx_lut[dist - 3];
     __m256i ret_vec;
     /* While technically we only need to read 4 or 8 bytes into this vector register for a lot of cases, GCC is
@@ -96,7 +98,7 @@ static inline chunk_t halfchunk2whole(halfchunk_t *chunk) {
     return lasx_zext_128(*chunk);
 }
 
-static inline halfchunk_t GET_HALFCHUNK_MAG(uint8_t *buf, uint32_t *chunk_rem, uint32_t dist) {
+static inline halfchunk_t GET_HALFCHUNK_MAG(uint8_t *buf, size_t *chunk_rem, size_t dist) {
     lut_rem_pair lut_rem = perm_idx_lut[dist - 3];
     __m128i perm_vec, ret_vec;
     __msan_unpoison(buf + dist, 16 - dist);
